requirements.txt
langchain>=0.3.25
langchain-core>=0.3.61
langchain-openai>=0.3.18
langchain-anthropic>=0.3.13
python-dotenv>=1.1.0
langchain-community>=0.3.24
duckduckgo-search>=8.0.2
wikipedia>=1.4.0
pyyaml>=6.0.2
anthropic>=0.52.0
langgraph>=0.4.5
pydantic
numexpr>=2.10.2